\name{CpDyna-package}
\alias{CpDyna-package}
\alias{CpDyna}
\docType{package}
\title{
 Clustering and multiple change points detection in dynamic graphs 
}
\description{
  This package is dedicated to dynamic graphs. It implements a model based approach that performs clusering and multiple change points  detection in dynamic graphs. 
}
\details{
  The most important function is called ModFit, it takes a graph as an argument (among other parameters) and perform a clustering as well as multiple change point detection.
}
\author{
Maintainers: Margot Selosse <margot.selosse@gmail.com>, Marco Corneli <Marco.CORNELI@unice.fr>.  
}
%\references{
%  This optional section can contain literature or other references for
%  background information.
%}
\keyword{ dynamic graphs, clustering, change points }
%\seealso{
%  Optional links to other man pages
%}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
