\name{Gnu}
\docType{data}
\alias{Gnu}
\title{Simulated dynamic graph}
\description{
 A toy dataset for running simple examples.
}
\usage{data(Gnu)}

\format{ A matrix with 3 rows and 1000 columns that represents 1000 undirected interactions between 40 nodes in a dynamic graph. The interaction times on the first row are unique, source nodes are stored in the second row and destination nodes in the third row.
}
\keyword{dataset; simulation}