\name{stations_df}
\docType{data}
\alias{stations_df}
\title{Santander stations}
\description{
More information about the Santander bike stations of London (see also ?wtab).
}
\usage{data(stations_df)}

\format{ 
A data.frame with 768 rows and 13 columns.

	\describe{
		\item{id}{}
		\item{name}{}
		\item{terminalName}{}
		\item{lat}{}
		\item{long}{}
		\item{installed}{}
		\item{locked}{}
		\item{installDate}{}
		\item{removalDate}{}
		\item{temporary}{}
		\item{nbBikes}{}
		\item{nbEmptyDocks}{}
		\item{nbDocks}{}
	}
}
\keyword{datasets; london; bike-sharing}